refactor: CLI structure and command handling
- Removed the CLI command definitions and related logic from `src/main.rs`, moving them to the `chrome_devtools_cli` crate.
- Updated the main function to utilize the new command structure and error handling.
- Simplified the command execution flow by integrating the `run` function from the `chrome_devtools_cli` crate.
- Enhanced the test suite in `executor_tests.rs` to ensure `known_args` remains in sync with the CLI command definitions, improving maintainability and reducing potential discrepancies.
